requirements.txt
install 
flask 
mysql-connector-python 
flask-sqlalchemy 
python-dotenv